# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de ejemplo con Clientes, Productos y Facturas",
        "title": "Microservicio Go + GORM + JWT",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/clientes": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Obtiene todos los clientes (requiere JWT)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clientes"
                ],
                "summary": "Lista clientes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Cliente"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Crea un nuevo cliente (requiere JWT)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clientes"
                ],
                "summary": "Crea un cliente",
                "parameters": [
                    {
                        "description": "Datos del cliente",
                        "name": "cliente",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Cliente"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Cliente"
                        }
                    }
                }
            }
        },
        "/facturas": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Lista todas las facturas y sus relaciones (requiere JWT)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facturas"
                ],
                "summary": "Listar facturas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Factura"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Crea una factura a partir de un cliente y productos (requiere JWT)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facturas"
                ],
                "summary": "Crear una factura",
                "parameters": [
                    {
                        "description": "Datos para crear factura",
                        "name": "factura",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.FacturaInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Factura"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Retorna un JWT válido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Credenciales",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/productos": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retorna todos los productos guardados (requiere JWT)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Listar productos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Producto"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Crea un nuevo producto en la base de datos (requiere JWT)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un producto",
                "parameters": [
                    {
                        "description": "Datos del producto",
                        "name": "producto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Producto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Producto"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Crea un nuevo usuario con email y password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Registro de usuario",
                "parameters": [
                    {
                        "description": "Datos de registro",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RegisterInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.FacturaInput": {
            "type": "object",
            "properties": {
                "cliente_id": {
                    "type": "integer"
                },
                "producto_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "handlers.LoginInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handlers.RegisterInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.Cliente": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
        "models.Factura": {
            "type": "object",
            "properties": {
                "cliente": {
                    "$ref": "#/definitions/models.Cliente"
                },
                "cliente_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "productos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Producto"
                    }
                },
                "total": {
                    "type": "number"
                }
            }
        },
        "models.Producto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                },
                "precio": {
                    "type": "number"
                }
            }
        }
    }
}